[TvShow] Move files widget to UI directory
