+++ src/ui/tv_show/TvShowFilesWidget.ui
